requirements.txt
fastapi==0.112.0
uvicorn[standard]==0.30.5
requests==2.32.3
python-dotenv==1.0.1
aiofiles==24.1.0
Jinja2==3.1.4
google-generativeai==0.7.2
pydantic-settings==2.4.0
